use mapped request:proposal from requests table
