+++ api/utils/eth.js
@@ -19,10 +19,12 @@ function checkConnection() {
 
 function getContracts() {
   const provider = new ethers.providers.JsonRpcProvider(rpcEndpoint);
-  const gatekeeper = new ethers.Contract(gatekeeperAddress, Gatekeeper.abi, provider);
-  const tokenCapacitor = new ethers.Contract(tokenCapacitorAddress, TokenCapacitor.abi, provider);
+  const signer = provider.getSigner();
+  const gatekeeper = new ethers.Contract(gatekeeperAddress, Gatekeeper.abi, signer);
+  const tokenCapacitor = new ethers.Contract(tokenCapacitorAddress, TokenCapacitor.abi, signer);
 
   return {
+    signer,
     provider,
     gatekeeper,
     tokenCapacitor,
